webapp/requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
Django==1.11.1
django-admin-easy==0.3.2
django-jet==1.0.6
django-modeladmin-reorder==0.2
django-rosetta==0.7.13
idna==2.5
microsofttranslator==0.8
numpy==1.13.1
packaging==16.8
pandas==0.20.3
polib==1.0.8
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.1
six==1.10.0
urllib3==1.21.1
xlrd==1.0.0
